Rename makecode projects more consistently when they are not the active project
Extract renameProject to project-utils so it can be shared between
store and storage. The shared version updates both header.name and
pxt.json name, and handles missing headers gracefully.
